[TG-TECH][refactor] scenarios folder structure
