--- src/app/components/scenario-wizard/index.js
@@ -1,1 +1,0 @@
-export * from './scenario-wizard.component';
