+++ src/app/scenarios/index.js
@@ -1,5 +1,7 @@
 export * from './scenarios-list/scenario-list.component.js';
-export * from './edit-scenario-editor/edit-scenario-editor.component';
-export * from './edit-scenario-wizard/edit-scenario-wizard.component';
-export * from './create-scenario-editor/create-scenario-editor.component';
-export * from './create-scenario-wizard/create-scenario-wizard.component';
+
+export * from './scenario-details/editor/edit-scenario-editor/edit-scenario-editor.component';
+export * from './scenario-details/editor/create-scenario-editor/create-scenario-editor.component';
+
+export * from './scenario-details/wizard/edit-scenario-wizard/edit-scenario-wizard.component';
+export * from './scenario-details/wizard/create-scenario-wizard/create-scenario-wizard.component';
