+++ src/app/scenarios/scenario-details/editor/create-scenario-editor/create-scenario-editor.component.js
@@ -3,11 +3,11 @@ import {RouteParams, Router, RouterLink} from 'angular2/router';
 import template from './create-scenario-editor.html';
 import style from './scenario-editor.scss';
 
-import {ScenarioDetailsComponent} from '../scenario-details/scenario-details.component';
+import {ScenarioDetailsComponent} from '../../scenario-details.component';
 import {CreateScenarioWizardComponent} from
-    '../create-scenario-wizard/create-scenario-wizard.component';
+    '../../wizard/create-scenario-wizard/create-scenario-wizard.component';
 
-import {ScenarioService} from '../Scenario.service.js';
+import {ScenarioService} from '../../../shared/Scenario.service.js';
 import {ScenarioEditor} from '../scenario-editor/scenario-editor.component.js';
 
 @Component({
