+++ src/app/scenarios/scenario-details/editor/create-scenario-editor/create-scenario-editor.html
