+++ src/app/scenarios/scenario-details/editor/create-scenario-editor/scenario-editor.scss
