+++ src/app/scenarios/scenario-details/editor/scenario-editor/code-mirror-events.const.js
