+++ src/app/scenarios/scenario-details/wizard/create-scenario-wizard/create-scenario-wizard.html
