+++ src/app/scenarios/scenario-details/wizard/scenario-entities/Condition.js
