+++ src/app/scenarios/scenario-details/wizard/scenario-entities/Scenario.js
