+++ src/app/scenarios/scenario-details/wizard/scenario-entities/logical-operators.const.js
