+++ src/app/scenarios/scenario-details/wizard/scenario-wizard/scenario-wizard.component.js
@@ -2,7 +2,7 @@ import {EventEmitter} from 'angular2/core';
 import {Component} from 'angular2/core';
 import template from './scenario-wizard.html';
 import style from './scenario-wizard.scss';
-import LOGICAL_OPERATORS from '../../scenarios/scenario-entities/logical-operators.const.js';
+import LOGICAL_OPERATORS from '../scenario-entities/logical-operators.const.js';
 const selector = 'scenario-wizard';
 
 @Component({
