+++ src/app/scenarios/scenario-details/wizard/scenario-wizard/scenario-wizard.component.test.js
@@ -1,4 +1,4 @@
-import {ScenarioWizardComponent} from './scenario-wizard.component';
+import {ScenarioWizardComponent} from './scenario-wizard.component.js';
 
 describe('ScenarioWizardComponent', () => {
     let sut;
