+++ src/app/scenarios/scenarios-list/assets/empty-scenarios-list.png
